fix: save() no longer mutates live data, set_speed uses Fraction, format_duration handles hours
save() now operates on a deep copy, preventing data corruption on failed writes
save() only warns on errors, not all validation issues (reduces noise from 836 to 9 warnings)
set_speed stores scalar as Fraction string consistent with scalar setter
format_duration outputs H:MM:SS.ff for durations over 1 hour
DropShadow.enabled property added
snap_to_grid produces integer start times

+++ src/camtasia/operations/layout.py
@@ -78,4 +78,4 @@ def snap_to_grid(track: Track, grid_seconds: float = 1.0) -> None:
         raise ValueError(f'Grid must be positive, got {grid_seconds}')
     for m in track._data.get('medias', []):
         start = m.get('start', 0)
-        m['start'] = max(0, round(start / grid_ticks) * grid_ticks)
+        m['start'] = max(0, int(round(start / grid_ticks) * grid_ticks))
